// src/models/IoK8sApimachineryPkgApisMetaV1StatusDetails/types.ts
import { IoK8sApimachineryPkgApisMetaV1StatusCauseModelOptions as IoK8sApimachineryPkgApisMetaV1StatusCause } from '../IoK8sApimachineryPkgApisMetaV1StatusCause';

export interface IoK8sApimachineryPkgApisMetaV1StatusDetailsModelOptions {
  causes?: Array<IoK8sApimachineryPkgApisMetaV1StatusCause>;
  group?: string;
  kind?: string;
  name?: string;
  retryAfterSeconds?: number;
  uid?: string;
}